This also produces an unused element warning
I think it's a bit futile to try and reverse engineer these optimizations. Compiler optimizations aren't exactly interpretable, whether it is for a CPU or an FPGA.

My conclusion is that we should do three more or less obvious things:
(1) Test the code to ensure it's actually correct.
(2) Test the code to ensure it runs at a particular speed.
(3) Turn off all optimizations or use the KEEP attribute to ensure something isn't touched (https://forums.xilinx.com/t5/Synthesis/KEEP-attribute-in-Verilog-amp-Timing-Optimization/td-p/241792)